requirements.txt
fastapi
uvicorn
paho-mqtt
websockets
jinja2
python-multipart
certifi
python-dotenv
opencv-python-headless
face-recognition
numpy
requests
cmake
dlib
